Continued work on the settings system, and paved the way for a new contact system
Most of the changes had to do with implementing the new Contact class in various locations., mainly implementing the "Self User" contact.  Basically the Contact class contains info about each other client and can store the display name, IP address, and a unique identifier that can be used to determine if somebody else has changed their name or and IP address change occurred for that contact.