Dynamic contacts system implemented, plus a bunch of objects and variables within objects changed to private
Added a new contacts system.  When you messages somebody for the first time, information is exchanged between the two peers to determine who is there, and then the name is stored, along with the IP address and a randomly generated ID for the computer.  The ID is used to determine if somebody changed their display name, but can still be mapped to the same IP address.  It is also used to prevent multiple contact entries of the same name, which can prevent errors and display name spoofing.  Other than that, most of the variables and objects within objects were changed to private.