Functioning Test Server
We not have a functioning test server that sends your messages back to you.  For some reason it has issues running on the same machine like I would normally do with another client.  However, it works perfectly fin on another machine or a virtual machine.  Since it's for testing, I'm not too worried about it.