TX and RX modules have been deleted.
These modules have been removed.  Functionality has changed over to MessagingControlModule and the ConnectionManager.  The MessagingControlModule now just handles data movement that is specific to the program itself.  The ConnectionManager is in charge of initiating all of the connections, and accepting new ones at a lower level.  The ConnectionManager also hands new connections off to different ports, so that way we aren't just hogging one port.  This system is much more extensible than the previous one and need be done about it.  This was also a crucial step to enabling multiple chat sessions at a time.