ConnectionManager is now functional.  MessageControlModule absorbed some duties.
The MessagingControlModule absorbed some of the roles that the RX and TX modules left behind.  The MessagingControlModule now just handles data movement that is specific to the program itself.  The ConnectionManager is in charge of initiating all of the connections, and accepting new ones at a lower level.  The ConnectionManager also hands new connections off to different ports dynamically, so that way we aren't just hogging two ports.  This system is much more flexible than the previous one and is needed to be done.  This was also a crucial step to enabling multiple chat sessions at a time.